tests and class comments
