+++ src/main/java/ua/tc/marketplace/facade/AdFacade.java
@@ -7,6 +7,13 @@ import ua.tc.marketplace.model.dto.ad.AdDto;
 import ua.tc.marketplace.model.dto.ad.CreateAdDto;
 import ua.tc.marketplace.model.dto.ad.UpdateAdDto;
 
+/**
+ * AdFacade defines the interface for operations related to ads within the system.
+ *
+ * <p>This interface includes methods for querying, creating, updating, and deleting ads. It
+ * provides a high-level abstraction for ad management, allowing for interaction with ad entities
+ * through data transfer objects (DTOs).
+ */
 public interface AdFacade {
   Page<AdDto> findAll(Map<String, String> filterCriteria, Pageable pageable);
