+++ src/main/java/ua/tc/marketplace/util/ad_filtering/filters/AttributeFilterSpecification.java
@@ -9,6 +9,13 @@ import ua.tc.marketplace.model.entity.Ad;
 import ua.tc.marketplace.model.entity.AdAttribute;
 import ua.tc.marketplace.util.ad_filtering.FilterSpecification;
 
+/**
+ * AttributeFilterSpecification is a component that implements the FilterSpecification interface for the Ad entity.
+ *
+ * <p>This class provides a JPA Specification for filtering ads based on specific attributes.
+ * It creates a query that matches ads where the attribute name and value are equal to the specified criteria.
+ */
+
 @Component("attribute")
 public class AttributeFilterSpecification implements FilterSpecification<Ad> {
