+++ src/main/java/ua/tc/marketplace/util/ad_filtering/filters/AuthorIdFilterSpecification.java
@@ -6,6 +6,13 @@ import org.springframework.stereotype.Component;
 import ua.tc.marketplace.model.entity.Ad;
 import ua.tc.marketplace.util.ad_filtering.FilterSpecification;
 
+/**
+ * AuthorIdFilterSpecification is a component that implements the FilterSpecification interface for
+ * the Ad entity.
+ *
+ * <p>This class provides a JPA Specification for filtering ads based on the author's ID. It creates
+ * a query that matches ads where the author ID is equal to the specified value.
+ */
 @Component("authorId")
 public class AuthorIdFilterSpecification implements FilterSpecification<Ad> {
