+++ src/main/java/ua/tc/marketplace/util/ad_filtering/filters/PriceMinFilterSpecification.java
@@ -7,6 +7,13 @@ import org.springframework.stereotype.Component;
 import ua.tc.marketplace.model.entity.Ad;
 import ua.tc.marketplace.util.ad_filtering.FilterSpecification;
 
+/**
+ * PriceMinFilterSpecification is a component that implements the FilterSpecification interface for
+ * the Ad entity.
+ *
+ * <p>This class is used to create a JPA Specification for filtering ads based on a minimum price.
+ * It generates a query that matches ads with a price greater than or equal to the specified value.
+ */
 @Component("priceMin")
 public class PriceMinFilterSpecification implements FilterSpecification<Ad> {
